created assertions for postgrest client test
